+++ NEWS.md
@@ -1,5 +1,6 @@
 # scda 0.1.1.9000
 
+* Added the `.onLoad` message with possible guidance to install an scda database package.
 * Updated R version requirement to >= 3.6.
 * Add `synthetic_cdisc_dataset` function to directly pull out a single dataset (which can be used in `DatasetConnector` in teal).